Added support for the third dimension.
Added support to zoom out and in meshes that is being drawn. Also added support for uniform variables in the shader programs that are being loaded. Added empty Transformation class that has to be implemented next to move code that handles Translations, Rotations and Scaling of things that should be drawn.